+++ wasatch/SpectrometerState.py
@@ -118,9 +118,6 @@ class SpectrometerState:
         self.graph_alternating_pixels = False
         self.swap_alternating_pixels = False
 
-        # Batch Collection
-        self.free_running_mode = True
-
         # mechanical articulation (e.g. Sandbox optics); this is currently 
         # treated as an integer with discrete steps
         self.position = 0